Deploy the geoapp to a Docker container in a Vagrant.
The vagrant file will create a virtualbox.  Within that virtualbox, a docker container is created for the geoapp.  Any docker containers within the virtualbox can be shown in a demo web page each under their own subpath.  The taxi data can be downloaded an installed.

The Vagrant box has running instances of Mongo and Postgres that can be shared by multiple docker containers.

This is an initial checkin that still needs more work.