Use triangles not sprites for point rendering.
Apparently, the true key to the MacBook speed issue is using triangles for rendering points rather than GL_POINTS (the 'sprite' option).  This means 1/3 as many points can be loaded, and, in the ideal case, we would measure the performance and pick the sprites when they work to be faster and use less memory when it works as well.  For now, just switch back to triangles.