Add instagram controls and points on map.
The controls are logged and routed.

Refactored how data is loaded to make it easier to add new data sets.  I could see doing more to allow parallel requests of the same data.

Refactored how maps are drawn so each set of data is in a separate layer and can have different rendering and animation behavior.