Fix the None range for maps.
There were some issues I discovered with style checks that allowed me to hunt down the issue with the 'None' range of the animation.

Fundamentally, if you explicitly set one of the data sets to a zero-length range, it will certainly use the range of the other data set.  If they are both non-zero length, then which layer gets picked for the range is uncertain.  This will be fixed when I add the ability to turn off the different data sets.