Highlight instagram points.
Show instagram photos and captions when you hover over them on the map or table.  We probably should make it so you can either move to a selected table entry or make the overlay sticky.

If the instagram point is not on the visible viewport, the overlay contains an arrow that points in the direction you need to move to see the point.

Perhaps the overlay should avoid the results table if it was created from the
results table.

Also, switched to building from the non-minified version of geojs.  We minify it when building, so the net result is the same, except that this allows chrome to show the unminified source.

Adjusted text queries to allow negation of quoted strings to not affect initial lexeme queries.

Typing enter on a text field will automatically update that control area.

Updated the geojs reference to match the new master.