Fix a variety of small issues.
Animation opacity is based on static data opacity.

Even though we check if an image can load, it still would sometimes take a while to render.  Make sure it is obvious which will happen.

Added target _blank to the instagram overlays.

The time in the instagram database appears to be Eastern Time.  I'm certain there are daylight savings issues.