Set location buttons from config settings.
Update to latest geojs so that we don't have to check if a feature is visible before handling mouse events.

Fixed an issue with the Filter button not highlighting when the 'Use Pickup Date Range' checkbox was toggled.

Fixed an issue where when you updated the display, sometimes the animation would be a strange state.  For instance, loading the page when the animation was stopped, then play the animation, pause it, and then update the display parameters.  The animation would resume even though it should have remained paused.