Instagram overlays can persist.
Instagram overlays can be sticky by clicking on the list or map.  Clicking a blank area on the map or the same point or list item will un-stick the overlay.  The overlay moves when the map is panned.

Fixed a minor bug where hovering over the instagram list table header would log an error to the console.

Update the speed test to handle other changes.

Updated to the latest geojs which removes the need for mousewheel.

Decreased the default verbosity.

Update Draper logging for the new elements.