Changes to the various dockerfiles and scripts.
Fix an issue with npm version.  Despite upgrading to the latest node version, npm was still woefully old and has to be updated explicitly.  The old version of npm was throwing errors on our demo machine's docker containers (but not on my virtualbox's docker containers).

Added a python3 docker for girder.

We should revisit our package.json and grunt file to make sure that we have partitioned the developer requirements from the production requirements.

Adjusted some css in the index page.