More graph features.
Add routing (navigation) for graphs.  If the data isn't available, indicate that we are waiting for data.  This includes the zoom range (the subchart brush area).

Added logging for graphs.  Refactored some of the logging to make it easier to dynamically add elements.

Added some internal events and triggers to make it easier to have producer and consumer processes.

Update graphs when data is loaded.

Moved graphs to upper right.

Add Stamen Terrain map.

Improve source maps of built javascript.