Refactored several controls.
Adjusted controls for rebinning.

Refactored animation controls.

If Instagram messages have been loaded, make them the default graph data set rather than temperature.

When we aren't using a database for a long time, disconnect from it.  Connection seems to only take 20 milliseconds, and this makes it easier for me to update databases in place, and for some resources to be freed.

Removed backbone from the package list, as we don't need to install it (it comes with Girder).