Added weekly and monthly options for graphs.
Removed d3 as a direct dependency, as we get it with Girder.

Changed the date graph axis to let it autocompute when rescaling the plot.  It sometimes has too dense tick labels, but that is better than too sparse.  I think I have to twiddle c3 or d3 to fix the problem.

Started working on the controls for a unified graph date range.

Better overlay tooltips.

Make instagram result list reactive to data visibility filtering.