Make all line graphs use the same date range, which is the date range of the loaded data.
All line graphs are linked.

Filter the points on the map based on the graph.

Change the opacity of scatter plot points based on date range.

The maximum zoom on a date graph is now 1 day.

Force d3 to use utc date so that the timezone of the user doesn't affect the data.

If you select a point on a graph, it zooms in to the larger of that day or the bin width of the graph.

Added Greater NYC to the locations (and removed Times Square).  Fix tooltips on location buttons.

Update d3 to the latest version.  This sort of undoes a previous change of not having a redundant d3 library.  The first version of d3 that is loaded is the one that is used.  Prior to this change, we were loading the grider one first, which meant that this version was ignored.  Now, we load this version of d3 first.

Changed the time scale formatting to look nicer at different resolutions.

Switched line graphs to step graphs.

Fix the tick density via some ugly monkey patching.