Don't reset the message list on updates, only on full replacements.  Maintain scroll position.
Reduce polling query load.

Remove support for pgdb, since we found that psycopg2 was faster.

Added a source field and a logging option.

Don't clear the overlay when data is updated (only when it is reloaded).

Fix an issue when the graph updates with tracking what point the user is over.