Better database connections.
Introduced the concept of a client to a rest end point.  If the client requeries before a preceding query is finished, the preceding query is cancelled.

A connection pool has been added so that multiple clients can make simultaneous queries, and database connections are kept open for a while for quicker connection.  If no client has used a connection for a while, it is closed.  If more clients make simultaneous queries than the pool is large, extra connections are opened as needed (up to Postgres's limit).  If a query fails for a reason (such as a temporary network glitch), a new connection is attempted rather than using one in the pool.

Initial work to support newer version of geojs.  Updated to the latest version that isn't broken for this app.

Added a grid tile service to make testing geojs parallel projection easier.