Add the ability to highlight recent messages.
Ideally, for time-based highlighting, we would adjust the points at least every 1/255 of the smaller of the recent and less-recent time intervals so that points would fade out on their own, rather than relying on new data coming in and causing a redraw.

Fixed a style on the message table for short dates.

Also, prepare for when geojs has parallel projection and discrete zoom, and fix a minor issue with a realtime ingest script.