Update to the latest geojs.
This at least partially addresses issue #45.  I can clearly see the jitter with the older version of geojs when zoomed in such that we exceed the tile zoom level 18 (using a large window, this is functionally zoom level 20.5 or 21).  This jitter is gone even when I allow zoom levels of 21 with the new version of geojs.