Added a work around for not having the ability to query for recent data.
This keeps track of urls that have been seen, and only reports distinct new urls.  To save memory (in case of many clients), it only saves hashes of the urls.  It will periodically redo the original query so as to reduce the memory footprint.  While polling, only recent data is checked for (this defaults to 1/2 hour but is configurable).  If this is set to a large value, then even old data will be discovered, but there will be substantially more load on the elasticsearch server and more load on the Minerva server.

When we are in a full streaming query and only looking at a small window (for instance, full Twitter with only looking at the last 10 minutes), this will be abusive to the elasticsearch database, but that may be unavoidable due to the data volume.