Basic support for mobile devices.
If the device has a small screen or is in portrait orientation, we attempt to set a scale factor that will allow all controls to be reachable, even if they are too small for easy use.  We handle pinch and pan touch events via the hammerjs library.

This works well on my touch-screen laptop and poorly on my Samsung phone.  We need to test it on something like a iPad.