More tuning for elasticsearch.
Indicate the state of polling so that you can tell the web page is still live.

Fix how text searches are scored in elasticsearch.

Restart polled queries if the server is unreachable for some reason.

Improve handling elasticsearch timeouts.

Improve documentation on elasticsearch configuration.

Fix an exception that was thrown when the server restarted and then a poll request was made from the client.

Updated geojs to the latest master.